requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argcomplete==1.10.3
asttokens==2.4.1
beautifulsoup4==4.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.6.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
comm==0.2.2
compressed_rtf==1.0.6
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
docx2txt==0.8
ebcdic==1.1.1
email_validator==2.2.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
exceptiongroup==1.2.1
executing==2.0.1
extract-msg==0.28.7
fastapi==0.111.0
fastapi-cli==0.0.4
fonttools==4.53.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
IMAPClient==2.1.0
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
murmurhash==1.0.10
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.25.0
nvidia-nccl-cu12==2.22.3
olefile==0.47
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdfminer.six==20191110
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
preshed==3.0.9
prompt_toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.1.2
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
shellingham==1.5.4
six==1.12.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
textract==1.6.5
thinc==8.2.4
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.22.0
wcwidth==0.2.13
weasel==0.4.1
websockets==12.0
wrapt==1.16.0
xgboost==2.1.0
xlrd==1.2.0
XlsxWriter==3.2.0
